Move validator generation to TerraFormFieldHelper
